Add modded content: tracks
This pushes in a bunch tracks I've made with RallySportED, and a PHP
script to generate a webpage listing those tracks.

Other people's tracks can be appended to the database also, though I'm
at present not aware of any. The JSON database already includes a field
for specifying the track's author, but at the moment the PHP script doesn't
spit that out onto the page (since there's just one author). Once multiple
authors are present in the JSON, the PHP should be modified accordingly to
inform the viewer.